requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2020.12.5        # via requests
chardet==4.0.0            # via requests
click==7.1.2              # via flask
ddtrace==0.48.0           # via -r requirements.in
flask==1.1.2              # via -r requirements.in
gevent==21.1.2            # via -r requirements.in
greenlet==1.0.0           # via gevent
gunicorn==20.1.0          # via -r requirements.in
idna==2.10                # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.3            # via flask
markupsafe==1.1.1         # via jinja2
protobuf==3.15.8          # via ddtrace
py-healthcheck==1.10.1    # via -r requirements.in
python-nomad==1.2.1       # via -r requirements.in
requests==2.25.1          # via python-nomad
six==1.15.0               # via protobuf, py-healthcheck, tenacity
tenacity==7.0.0           # via ddtrace
urllib3==1.26.4           # via requests
werkzeug==1.0.1           # via flask
zope.event==4.5.0         # via gevent
zope.interface==5.3.0     # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
